+++ lib/asset_sync/storage.rb
@@ -12,6 +12,14 @@ module AssetSync
     end
 
     def connection
+      begin
+        # TODO: make a config option for the provider requirement that should be dynamically loaded
+        # require 'fog/storage' unless defined?(Fog::Storage)
+        require 'fog/aws/storage'
+      rescue LoadError => e
+        log('You must include your fog provider gem in your Gemfile')
+        raise e
+      end
       @connection ||= Fog::Storage.new(self.config.fog_options)
     end